# docs/generated/swagger.yaml
openapi: 3.1.0
info:
  title: 1BIS Backend API
  version: 1.0.0
  description: Welcome to the 1BIS API documentation. This is a RESTful API that
    allows you to interact with the 1BIS backend services.
  contact:
    name: 1BIS Support
    email: support@1bis.com
    url: https://1bis.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://1bis.com/terms
servers:
  - url: /api
    description: Development server
  - url: https://api.1bis.com/api
    description: Production server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Authorization header using the Bearer scheme. Example:
        'Authorization: Bearer {token}'"
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key for authentication
...
paths:
...
  /api/transaction/{id}:
    get:
      summary: Get transaction by ID
      description: Retrieve a specific transaction by its unique identifier with
        optional field selection
      tags:
        - Transaction
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          description: Transaction ID (MongoDB ObjectId format)
          example: 507f1f77bcf86cd799439011
        - in: query
          name: fields
          required: false
          schema:
            type: string
          description: Comma-separated list of fields to include (supports nested fields
            with dot notation)
          example: id,name,description,type
      responses:
        "200":
          description: Transaction retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          transaction:
                            $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    patch:
      summary: Update transaction
      description: Update transaction data by ID (partial update)
      tags:
        - Transaction
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          description: Transaction ID (MongoDB ObjectId format)
          example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  minLength: 1
                  description: Transaction name
                  example: Updated Email Transaction
                description:
                  type: string
                  description: Transaction description
                  example: Updated description for the transaction
                type:
                  type: string
                  enum:
                    - email
                    - sms
                    - push
                    - form
                  description: Transaction type for categorization
                  example: email
                isDeleted:
                  type: boolean
                  description: Soft delete flag
                  example: false
      responses:
        "200":
          description: Transaction updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          transaction:
                            $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete transaction
      description: Permanently delete a transaction by ID
      tags:
        - Transaction
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            pattern: ^[0-9a-fA-F]{24}$
          description: Transaction ID (MongoDB ObjectId format)
          example: 507f1f77bcf86cd799439011
      responses:
        "200":
          description: Transaction deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        description: Empty object for successful deletion
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/transaction:
    get:
      summary: Get all transactions
      description: Retrieve transactions with advanced filtering, pagination, sorting,
        field selection, and optional grouping
      tags:
        - Transaction
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
          example: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of records per page
          example: 10
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum:
              - asc
              - desc
            default: desc
          description: Sort order for results
          example: desc
        - in: query
          name: sort
          required: false
          schema:
            type: string
          description: Field to sort by or JSON object for multi-field sorting
          example: createdAt
        - in: query
          name: fields
          required: false
          schema:
            type: string
          description: Comma-separated list of fields to include (supports dot notation)
          example: id,name,description,type
        - in: query
          name: query
          required: false
          schema:
            type: string
          description: Search query to filter by name or description
          example: welcome email
        - in: query
          name: filter
          required: false
          schema:
            type: string
          description: JSON array of filter objects for advanced filtering
          example: '[{"type":"email"},{"isDeleted":false}]'
        - in: query
          name: groupBy
          required: false
          schema:
            type: string
          description: Group results by a field name
          example: type
        - in: query
          name: document
          required: false
          schema:
            type: string
            enum:
              - "true"
          description: Include transaction documents in response
        - in: query
          name: pagination
          required: false
          schema:
            type: string
            enum:
              - "true"
          description: Include pagination metadata in response
        - in: query
          name: count
          required: false
          schema:
            type: string
            enum:
              - "true"
          description: Include total count in response
      responses:
        "200":
          description: Templates retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          transactions:
                            type: array
                            items:
                              $ref: "#/components/schemas/Transaction"
                            description: Present when document="true" and no groupBy
                          groups:
                            type: object
                            additionalProperties:
                              type: array
                              items:
                                $ref: "#/components/schemas/Transaction"
                            description: Present when groupBy is used and document="true"
                          count:
                            type: integer
                            description: Present when count="true"
                          pagination:
                            $ref: "#/components/schemas/Pagination"
                            description: Present when pagination="true"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create new transaction
      description: Create a new transaction with the provided data
      tags:
        - Transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  description: Transaction name
                  example: Email Welcome Transaction
                description:
                  type: string
                  description: Transaction description
                  example: Welcome email transaction for new users
                type:
                  type: string
                  enum:
                    - email
                    - sms
                    - push
                    - form
                  description: Transaction type for categorization
                  example: email
                isDeleted:
                  type: boolean
                  description: Soft delete flag
                  default: false
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                description:
                  type: string
                type:
                  type: string
                isDeleted:
                  type: boolean
          multipart/form-data:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                description:
                  type: string
                type:
                  type: string
                isDeleted:
                  type: boolean
      responses:
        "201":
          description: Transaction created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          transaction:
                            $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalServerError"
...